Rename serializer.cpp to less ambiguous name
