+++ rir/src/recording_serialization.h
